Two main new things here: PathExplorer and Outputter
PathExplorer is a way to just look at paths in the graph between pairs of
nodes.  This was originally motivated by a desire to see why the models I was
making weren't working, but it could also have some application to building
simpler feature matrices.

Outputter is a refactoring of all of the output methods for the PRA code,
removing the need to have a separate OutputTranslator.  Now the outputter
knows if it has Dictionaries, and can just output one file containing the best
strings it can make (i.e., using node names instead of ids, etc.).  This
should also make the output files a lot easier to work with, and should cut
the output size roughly in half.

There's also some reformatting of a few classes in here, as I'm moving to
using 2 spaces instead of 4.