Main addition: Testing for the GraphCreator
I'm about to improve the functionality of the GraphCreator, and I wanted to add some tests before I
did that.  Making GraphCreator testable required adding some functionality to FileUtil, so I did
that.  And while I was at it, I removed all static methods from FileUtil, which required changing
a few other files a bit, too.  And I finally added a makePair method to the Pair class, which has
been bugging me for a while with how verbose it is to create new Pairs.