Finished new spec file format, and fixed JVM exit
There used to be an odd string of error messages when PRA finished, because the GraphChi code
would catch an exception and print it out.  I removed that for InterruptedException, which happens
when the main code calls System.exit().  I also made the sbt console behave better by running PRA
in a forked process.  So now things should actually work as they're supposed to when the PRA code
terminates.