A lot of work on improving the spec file situation
You can now, from a single spec file and with running a single binary,
generate a synthetic dataset, create a graph from it, and run PRA.  And the
format for the spec files is more sane, allowing loads in the middle of the
file (mapFields in json4s is very nice!).

I still need to make it so you can create an embedding and densify the matrix
from the single spec file, plus a few other things.  But I'm getting very
close now.