Done with the spec rewrite, I think.
The biggest part of this commit is giving Driver.scala the ability to create
embeddings, similarity matrices, and denser graph matrices, and all of the
associated changes with that.  I also moved the functionality of
ExperimentExplorer into an option that can be passed in the spec file, and
removed a couple old main functions, so that now there are only two:
ExperimentRunner and ExperimentScorer.  I didn't write tests for all of the
new functionality, though...  It would take a lot of factories, and I thought
that was kind of obnoxious.

And there are a few minor cleanup items in here too, like switching from
scalax.io.Resource to my FileUtil, and switching from JavaConversions to
JavaConverters.