Some more fixes to the spec file code
I realized I had not yet migrated over the relation set stuff, so it wasn't
right.  I had to move the code from java to scala, so that I could use json4s
to pass in parameters.  That let me get rid of the Builder, which was nice.  I
also removed the obsolete GraphConfig, and made a few other minor changes I
noticed as I was updating the documentation.