More moving stuff around; close to done for now
This let me remove the whole PathFollowerFactory notion, which I'm happy with.
Test coverage has gone slightly down, but I added in a few tests on
FeatureGenerator to cover some of the things that I removed.

Also note that this commit totally changes the specification format for
"pra parameters" - until I bump the version number and update the documentation
with it, the code base will not match the documentation.  At least it should
tell you when you've specified something incorrectly.