The new feature extraction from subgraph is working
Mostly, anyway.  It runs, it works, but it doesn't handle negative examples in
the same way as the rest of the code.  Currently, negative examples need to be
supplied in the training and testing files for it to work right.  Either I can
just live with that and explicitly pick negative examples as part of a split
from now on, or I need to come up with a way to get negative examples at
training time and potential targets at test time.