Refactoring PathFinder to be a trait
This is so that I can implement a new PathFinder that uses a breadth first
search on an in-memory graph, instead of using GraphChi.  My hope is that this
will fix two issues:

(1) It paves the way for me to let each training instance have its own
associated small graph, which doesn't fit the GraphChi model very well.

(2) Assuming there's enough memory and it doesn't kill processing time, this
should let me fix coverage issues, when random walks don't find particular path
types that would be informative when a node has a very high degree.