A lot of (mostly minor) fixes, looking at profiling
I used a profiler to figure out what was making the code slow.  The biggest
thing was discovering that MALLET expects the input features to be sorted,
which I didn't know, so I fixed that, and saw pretty dramatic improvements.

Then I discovered that a lot of time was going to boxing in Index[PathType],
so I moved that to scala to remove the boxing.  In the process, I discovered
some other concurrency issues (that should have also existed in the java
version, but maybe the boxing time made it a non-issue?).

And I found that just concatenating strings is a bit faster than using string
interpolation, and when extracting features we produce a lot of strings, so I
removed string interpolation from the most frequented parts of the code (or at
least started to).

There may be some other things that I could fix after looking more at the
profiler, but I think I hit the biggest ones.