Added a "bad vector" to SimilarityMatrixCreator
It turns out that with several SVDs I've done on these graphs, a very large percentage of the
vectors end up being just about exactly the same.  The idea here was just to exclude all of those
when creating the similarity matrix.  It only kind of worked, and there's no good way to set the
bad vector without some hackish code.  So, if you want to use it, modify this code to set the bad
vector in a hackish way for whatever you're doing.