Added a special 'delete' keyword to the spec file
This lets you use a 'load' statement to get a set of parameters, then remove one or more of them
that you want to get rid of (when just overriding the parameter is not good enough).  This
shouldn't be used often, but it's really useful to avoid a lot of duplication in a few specific
circumstances.