MAJOR refactoring, moving Dataset and Outputter to scala
This was so that I can change the way Dataset works, allowing for individual
instances to have their own graphs.  Doing that will be a lot easier with this
code in scala.  Moving Dataset made me move Outputter, and both of those made
me touch a whole lot of other files.  But it is done.  Hopefully.  At least
all of the tests pass.  There are probably other things I can now simplify
because of these changes (like, not using java collections as much), but I'll
have to do that in another commit sometime later.