Implemented a format for instance-specific graphs
This should let me run tests where we run PRA with a few sentences and a
specific query.  I also added the dictionaries to the graph object, which
makes sense to me, because they're inherently a part of the graph, anyway.
This necessitated a few other changes in the API, which made me touch a lot of
files, but most of the changes are small.