Moved the node and edge dicts to the Graph
And this snowballed into touching pretty much every single file in the
codebase...

If you move the dictionaries into the graph, and each Instance in a Dataset
can be associated with a different graph, that means that passing around pairs
of Ints without saying what graph they belong to quite possibly wrong - a
matching Int pair might not even be defined with respect to the same graph.
So, then, each Instance needs to have a reference to the graph that it belongs
to, and we should pass around Instance objects instead of Int pairs.  And all
of this was just a big mess.

But, I think it's all done.  All of the tests are passing, and so hopefully
everything just works.  I haven't run the code yet, though...