Added a LexicalizePathType
This lets you include nodes in the path description, instead of just edge
sequences, which can be helpful in some circumstances.  It's not fully
implemented for PathFollowers yet, though, as I'm not anticipating using it
with the old PRA code, just with SFE, and it should work fine for SFE.

Adding this made me change the API for PathType slightly, which necessitated
making small changes to a bunch of files.  But the only major change here is
the addition of LexicalizedPathType.