Pulled the PPR computation out of PprNegativeExampleSelector
This was so that I could add different PPR implementations.  I want to push
toward just using the graph in memory whenever possible, while still allowing
GraphChi when it's needed.  So, this let me pull out the GraphChi PPR
implementation and add one that just does some random walks over the graph in
memory.

There are also a few improved comments and other minor things in this commit.