Added a "create feature matrix" PRA mode
This mode quits after constructing a feature matrix and before learning a
model.  The point is, if you only want to construct a feature matrix for use
in some other model, this will do it for you.  A better solution for this use
case would be to just create a SubgraphFeatureGenerator object and use it to
create your matrix, but setting that up is kind of a pain.