Starting some major refactoring of Driver
THIS IS PROBABLY A BUGGY COMMIT, EVEN THOUGH TESTS PASS

I'm changing the parameter names a bit, here, too, to make this conform more
to how other parts of the parameter specification work.

I was working on getting SGD-style training into the code, and decided I'd
refactor Driver, because it was a part of Driver that I wanted to make more
modular.  So now it is.