Experimenting with momentum for SGD training
My conclusion from the experiments I've done is that I need to do something smarter here.  I need
to test out a few different methods for doing SGD training (like AdaGrad, a few others) and see if
I can find something that works.  But I'm going to do that experimenting later, in a different
code base, and integrate it into this when it's done.