Updated the util library to version 1.2
1.2 of that library does two things: it moves FileUtil to scala, and it allows
for parallel file processing.

This should make a few things a bit faster, because we are no longer
converting to and from java objects in as many places.  But, more importantly,
it should let me read in very large files (like the SVO data) much more
efficiently.  Let's hope it actually works...