Beginning to add single-node feature extraction
That is, instead of just allowing feature extraction over node _pairs_ in a
graph, I want to also be able to extract features over _nodes_ in the graph.
This means I need the Instance class to have either one or two nodes, and the
Dataset to change accordingly.  Most of the code here will only be applicable
to the node pair case, but some of it will generalize.

This is an initial step in getting this all to work, by making Instance
abstract and Dataset take an Instance type parameter.  I also moved Dataset
and Instance into a data package.  These changes necessitated touching a whole
lot of code, which is why this commit is so big.  Next step: allow reading
Datasets of single nodes.  I'll probably improve the Split object situation in
the process, as I've been wanting to do that for a while.