Fixed dataset reading, so it's a lot cleaner
And I removed the split from PraConfig, which never should have been put in.
This let me remove the type parameter on PraConfig that I had to add when I
put a type parameter on Dataset.  So this is another one that touches a lot of
files, though the update is more minor on each one here.