Pushed the type parameters as far they could go
Now there are a bunch of abstract classes with the bulk of the SFE code that
take type parameters, with a few concrete classes that fill in the Instance
specific stuff for NodePairInstance.  All I need to do, I think, is add
parallel implementations for NodeInstance, and things should just work.
Hopefully...  There's also one type warning left that I need to figure out.