Provided implementations for Dataset[NodeInstance]
I got this to successfully run on a small example with a split of type "node".
 There are still one or two things to figure out to actually run a
classification experiment (such as leaving out the type edges in the graph, or
implementing some way to exclude them from a search, as well as generating
negative instances at training and test time), but for generating features
over nodes in a graph, the code now works, and you can define whatever feature
extractor you want over the extracted subgraphs.

I'm pretty happy with how this turned out; scala's type system is very nice.

I also fixed a compiler warning that was due to using Option[Operation[_ <:
Instance]].  I'm still not really sure what caused the warning, but removing
the Option fixed it...