Improvements to the Graph
Two main things in here:
- An API change for Dictionary, with the hope of greatly reducing memory usage
  for the graph dictionaries.  When you know that the node and edge
dictionaries aren't going to change, you can now specify that, and use a
dictionary object that should be faster and much more compact.  There were a
couple places in the code that assumed I could always just add an edge type to
the dictionary, so this change necessitated a few other changes.

- Trying to make RemoteGraph not block so much, by adding multiple connections
  to the graph server.