Added a binary file format for graphs
This cut my load time about in half for the Freebase graph (and that's after
previous recent improvement) - loading the graph is down to under 3 minutes,
from around 10 previously.

With this, I'm also bumping the version number and pushing the current version
to sonatype, so that another project I'm working on can use this easily.