Moved GraphCreator to using my pipeline code
This lets me include it as a Step in another project I'm working on. I
modified the Driver to also use the pipeline code, because it depends on the
GraphCreator, and that made the Driver slightly cleaner, I think.  Ideally, I
would move all of the other file creation dependencies in Driver to using the
pipeline code, but that will wait until I have sufficient motivation for it.

I haven't actually tested that this didn't break my experiment code, because I
don't have tests on the integration of some of these things (the unit tests
all pass).  So, hopefully everything still works...