Implemented a new kind of evaluation for SFE: SV? and ?VO queries
This is following common evaluation metrics in embedding methods for KBC.
It's not super efficient yet, and I need to fix the hacky implementation, but
it works.

As part of this, I updated some of the RelationMetadata API, which caused
errors in the MatrixPathFollowers.  I decided it was long past time to remove
those from the code, because they were a failed experiment, and there's no
point having that dead code take up maintanence time.