Got this actually working, with a more or less sane data structure
The previous commit was with minimal changes to how the processing was done,
so it used an incredibly redundant data structure that just wastes memory and
CPU.  This commit removes the inefficiency.  The code is still ~35% slower
than the current master branch, which is a little surprising to me.  I need to
do some profiling to figure out what's going on, because I thought this would
be faster.