Moving towards pipeline architecture for running stuff
In particular, I moved the previous hacky operation I wrote for Hanie into a
couple of Steps in my pipeline architecture.  This should let me make this
whole thing a lot more modular.  There's still more to do if I want to migrate
the whole ExperimentRunner / ExperimentScorer over to this, but at least this
is a start.  And I should be able to run experiments for Hanie with just a
little bit of work on top of this.

Also, this touches almost every single file because I finally moved to a real
logging framework.  I had previously made the poor decision of trying to
handle all logging in an object I passed around.  There were a few compelling
reasons for that initially, but expanding it to all logging statements was a
bad idea, and it led to a pretty ugly design.  I switched to ScalaLogging, and
removed the logging pieces from the Outputter.  This will eventually just make
the Outputter obsolete, which is a good thing.